+++ src/lib/clients/models/box-response.ts
@@ -1,0 +1,13 @@
+
+import { Box } from './box';
+
+/**
+ * Represents the HTTP API model for a response with a single box.
+ */
+export interface BoxResponse {
+
+    /**
+     * Gets or sets the requested box.
+     */
+    box: Box;
+}